@@ -9,12 +9,12 @@
     "url": "https://github.com/imjasonmiller/webgl_prototype"
   },
   "scripts": {
-    "test": "eslint src/*",
-    "build:db": "cross-env sh ./db.create.sh",
+    "test": "cross-env eslint --ext .js --ext .jsx src",
+    "db": "cross-env sh ./db.create.sh",
     "dev": "cross-env DEV=true node api/server.js",
-    "build:prod":
+    "prod:build":
       "cross-env rimraf ./build/* && webpack --config ./webpack/webpack.prod.js",
-    "start:prod": "cross-env node api/server.js"
+    "prod:start": "cross-env node api/server.js"
   },
   "author": "Jason Miller",
   "license": "ISC",
@@ -23,8 +23,8 @@
     "babel-loader": "^8.0.0-beta.0",
     "bcryptjs": "^2.4.3",
     "bunyan": "^1.8.12",
-    "compression-webpack-plugin": "^1.1.2",
-    "cross-env": "^5.1.1",
+    "compression-webpack-plugin": "^1.1.3",
+    "cross-env": "^5.1.3",
     "csrf": "^3.0.6",
     "dotenv": "^4.0.0",
     "file-loader": "^1.1.6",
@@ -34,10 +34,11 @@
     "koa-bodyparser": "^4.2.0",
     "koa-router": "^7.3.0",
     "koa-static": "^4.0.2",
-    "koa-webpack": "^1.0.0",
+    "koa-webpack": "^1.1.0",
     "lodash": "^4.17.4",
     "pg": "^7.4.0",
     "prop-types": "^15.6.0",
+    "raw-loader": "^0.5.1",
     "react": "^15.6.2",
     "react-dom": "^15.6.2",
     "react-hot-loader": "^3.1.3",
@@ -48,9 +49,9 @@
     "redux": "^3.7.2",
     "redux-saga": "^0.16.0",
     "rimraf": "^2.6.2",
-    "styled-components": "^2.3.0",
+    "styled-components": "^2.4.0",
     "three": "^0.89.0",
-    "uglifyjs-webpack-plugin": "^1.1.4",
+    "uglifyjs-webpack-plugin": "^1.1.5",
     "webpack": "^3.10.0",
     "webpack-bundle-analyzer": "^2.9.1",
     "webpack-hot-server-middleware": "^0.3.1",
@@ -61,7 +62,7 @@
     "@babel/preset-env": "^7.0.0-beta.35",
     "@babel/preset-react": "^7.0.0-beta.35",
     "@babel/preset-stage-0": "^7.0.0-beta.35",
-    "eslint": "^4.13.1",
+    "eslint": "^4.14.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-config-prettier": "^2.9.0",
     "eslint-import-resolver-webpack": "^0.8.3",
