@@ -17,7 +17,7 @@ const DB_PASS = process.env.DB_PASS ? `:${process.env.DB_PASS}` : ""
 // Logging
 const bunyan = require("bunyan")
 
-const log = bunyan.createLogger({ name: "dev server" })
+const log = bunyan.createLogger({ name: "Server" })
 
 const Koa = require("koa")
 const Router = require("koa-router")
@@ -52,11 +52,10 @@ if (DEV === "true") {
 
   const compiler = webpack(webpackConfig)
 
-  const devHot = koaWebpack({
+  const hotDev = koaWebpack({
     compiler,
     dev: {
-      publicPath:
-        webpackConfig.find(c => c.name === "client").output.publicPath || "/",
+      publicPath: "/",
       serverSideRender: true,
       stats: { colors: true },
     },
@@ -73,7 +72,7 @@ if (DEV === "true") {
     .use(koaBodyparser())
     .use(routes)
     .use(router.allowedMethods())
-    .use(devHot)
+    .use(hotDev)
     .use(hotServer)
 } else {
   // React server bundle, generated by Webpack
@@ -84,7 +83,7 @@ if (DEV === "true") {
     .use(koaBodyparser())
     .use(routes)
     .use(router.allowedMethods())
-    .use(server())
+    .use(server("lol"))
 }
 
 databaseConn(db)
