@@ -1,5 +1,5 @@
 export { default as ButtonStats } from "./ButtonStats"
 export { default as ButtonTool } from "./ButtonTool"
 
-export { default as Loader } from "./Loader"
 export { default as Modal } from "./Modal"
+export { default as Spinner } from "./Spinner"
