@@ -1,2 +1,2 @@
 export { default as CelestialBody } from "./CelestialBody"
-export { default as World } from "./World"
+export { default as Terrain } from "./Terrain"
